requirements.txt
django-allauth[mfa,saml,socialaccount,steam]==65.10.0
whitenoise
psycopg>=3.1
pgvector[django]
openai==1.82.0
python-dotenv==1.1.0
dj-database-url
tiktoken==0.9.0
PyPDF2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
pdfminer.six==20221105
pytest==8.4.1
responses>=0.24
freezegun
pytest-django
pytest-mock
responses[pytest]>=0.25
pytest-responses==0.5.1

